moved dict_check and str_dict_check from error_mngr to tools
