+++ src/bootstraparse/modules/error_mngr.py
@@ -83,33 +83,6 @@ def log_exception(exception, level="ERROR"):
         return
 
 
-def dict_check(dic, *args):
-    """
-    Helper function to test existence within a dictionary.
-    :param dic: The dictionary of dictionaries to check
-    :param args: Any number of keys to check the dictionaries for
-    :return: A table of Booleans for every key checked.
-    """
-    output = [False for _ in args]
-    for i in range(len(args)):
-        if args[i] in dic:
-            output[i] = True
-            dic = dic[args[i]]
-        else:
-            break
-    return output
-
-
-def str_dict_check(dic, *args):
-    """
-    Helper function to test existence within a dictionary and return a string of the values.
-    'arg[0]' : True or False
-
-    """
-    out = dict_check(dic, *args)
-    return "\n".join([f"{args[i]}: {out[i]}" for i in range(len(out))])
-
-
 class BootstraparseError(Exception):
     pass
